# apps/api/spec/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "acting-api",
    "description" : "연기 분석 플랫폼 API v2. Bearer access token을 사용합니다.",
    "version" : "0.1.0"
  },
  "paths" : {
...
    "/v2/practice-sessions" : {
      "get" : {
        "tags" : [ "v2-practice" ],
        "summary" : "List Sessions",
        "operationId" : "list_sessions_v2_practice_sessions_get",
        "responses" : {
          "200" : {
            "description" : "Successful Response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PracticeSessionListResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "HTTPBearer" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "v2-practice" ],
        "summary" : "Create Session",
        "operationId" : "create_session_v2_practice_sessions_post",
        "parameters" : [ {
          "name" : "X-Request-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ],
            "title" : "X-Request-Id"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PracticeSessionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful Response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PracticeSessionCreateResponse"
                }
              }
            }
          },
          "202" : {
            "description" : "Accepted",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PracticeSessionAcceptedResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Validation Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security" : [ {
          "HTTPBearer" : [ ]
        } ]
      }
    },
    "/v2/practice-sessions/{session_id}/analyze" : {
      "post" : {
        "tags" : [ "v2-practice" ],
        "summary" : "Reanalyze Session",
        "operationId" : "reanalyze_session_v2_practice_sessions__session_id__analyze_post",
        "parameters" : [ {
          "name" : "session_id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Session Id"
          }
        }, {
          "name" : "X-Request-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ],
            "title" : "X-Request-Id"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PracticeSessionCreateResponse"
                }
              }
            }
          },
          "202" : {
            "description" : "Accepted",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PracticeSessionAcceptedResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Validation Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security" : [ {
          "HTTPBearer" : [ ]
        } ]
      }
    },
...
    "/v2/practice-sessions/{session_id}" : {
      "get" : {
        "tags" : [ "v2-practice" ],
        "summary" : "Get Session",
        "operationId" : "get_session_v2_practice_sessions__session_id__get",
        "parameters" : [ {
          "name" : "session_id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Session Id"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PracticeSessionDetail"
                }
              }
            }
          },
          "422" : {
            "description" : "Validation Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security" : [ {
          "HTTPBearer" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "v2-practice" ],
        "summary" : "Delete Session",
        "operationId" : "delete_session_v2_practice_sessions__session_id__delete",
        "parameters" : [ {
          "name" : "session_id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Session Id"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "422" : {
            "description" : "Validation Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security" : [ {
          "HTTPBearer" : [ ]
        } ]
      }
    },
    "/v2/practice-sessions/{session_id}/status" : {
      "get" : {
        "tags" : [ "v2-practice" ],
        "summary" : "Get Session Status",
        "operationId" : "get_session_status_v2_practice_sessions__session_id__status_get",
        "parameters" : [ {
          "name" : "session_id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Session Id"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PracticeSessionStatusResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Validation Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security" : [ {
          "HTTPBearer" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "PracticeSessionRequest" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "upload_intent_id" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Upload Intent Id"
          },
          "situation" : {
            "type" : "string",
            "title" : "Situation"
          },
          "character_context" : {
            "type" : "string",
            "title" : "Character Context"
          },
          "goal" : {
            "type" : "string",
            "title" : "Goal"
          },
          "blockage_kind" : {
            "type" : "string",
            "enum" : [ "분석", "표현", "그 외" ],
            "title" : "Blockage Kind"
          },
          "sub_branch" : {
            "type" : "string",
            "enum" : [ "캐릭터 분석", "대사 분석", "감정", "움직임", "화술", "표정", "그 외" ],
            "title" : "Sub Branch"
          },
          "blockage_detail" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ],
            "title" : "Blockage Detail"
          },
          "continued_from" : {
            "anyOf" : [ {
              "type" : "string",
              "format" : "uuid"
            }, {
              "type" : "null"
            } ],
            "title" : "Continued From"
          }
        },
        "required" : [ "blockage_kind", "character_context", "goal", "situation", "sub_branch", "upload_intent_id" ],
        "title" : "PracticeSessionRequest"
      },
      "PracticeSessionCreateResponse" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "session_id" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Session Id"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "analyzing", "analyzed", "failed" ],
            "title" : "Status"
          },
          "summary_id" : {
            "anyOf" : [ {
              "type" : "string",
              "format" : "uuid"
            }, {
              "type" : "null"
            } ],
            "title" : "Summary Id"
          }
        },
        "required" : [ "session_id", "status" ],
        "title" : "PracticeSessionCreateResponse"
      },
      "PracticeSessionAcceptedResponse" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "session_id" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Session Id"
          },
          "status" : {
            "type" : "string",
            "const" : "analyzing",
            "title" : "Status"
          }
        },
        "required" : [ "session_id", "status" ],
        "title" : "PracticeSessionAcceptedResponse"
      },
...
      "HTTPValidationError" : {
        "type" : "object",
        "properties" : {
          "detail" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ValidationError"
            },
            "title" : "Detail"
          }
        },
        "title" : "HTTPValidationError"
      },
      "ValidationError" : {
        "type" : "object",
        "properties" : {
          "loc" : {
            "type" : "array",
            "items" : {
              "anyOf" : [ {
                "type" : "string"
              }, {
                "type" : "integer"
              } ]
            },
            "title" : "Location"
          },
          "msg" : {
            "type" : "string",
            "title" : "Message"
          },
          "type" : {
            "type" : "string",
            "title" : "Error Type"
          },
          "input" : {
            "title" : "Input"
          },
          "ctx" : {
            "type" : "object",
            "title" : "Context"
          }
        },
        "required" : [ "loc", "msg", "type" ],
        "title" : "ValidationError"
      },
...
      "PracticeSessionListItem" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "session_id" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Session Id"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "analyzing", "analyzed", "failed" ],
            "title" : "Status"
          },
          "situation" : {
            "type" : "string",
            "title" : "Situation"
          },
          "character_context" : {
            "type" : "string",
            "title" : "Character Context"
          },
          "goal" : {
            "type" : "string",
            "title" : "Goal"
          },
          "blockage_kind" : {
            "type" : "string",
            "enum" : [ "분석", "표현", "그 외" ],
            "title" : "Blockage Kind"
          },
          "sub_branch" : {
            "type" : "string",
            "title" : "Sub Branch"
          },
          "blockage_detail" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ],
            "title" : "Blockage Detail"
          },
          "continued_from" : {
            "anyOf" : [ {
              "type" : "string",
              "format" : "uuid"
            }, {
              "type" : "null"
            } ],
            "title" : "Continued From"
          },
          "created_at" : {
            "type" : "string",
            "format" : "date-time",
            "title" : "Created At"
          },
          "updated_at" : {
            "type" : "string",
            "format" : "date-time",
            "title" : "Updated At"
          }
        },
        "required" : [ "blockage_kind", "character_context", "created_at", "goal", "session_id", "situation", "status", "sub_branch", "updated_at" ],
        "title" : "PracticeSessionListItem"
      },
      "PracticeSessionListResponse" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "sessions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PracticeSessionListItem"
            },
            "title" : "Sessions"
          }
        },
        "required" : [ "sessions" ],
        "title" : "PracticeSessionListResponse"
      },
      "PracticeSessionDetail" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "session_id" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Session Id"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "analyzing", "analyzed", "failed" ],
            "title" : "Status"
          },
          "situation" : {
            "type" : "string",
            "title" : "Situation"
          },
          "character_context" : {
            "type" : "string",
            "title" : "Character Context"
          },
          "goal" : {
            "type" : "string",
            "title" : "Goal"
          },
          "blockage_kind" : {
            "type" : "string",
            "enum" : [ "분석", "표현", "그 외" ],
            "title" : "Blockage Kind"
          },
          "sub_branch" : {
            "type" : "string",
            "title" : "Sub Branch"
          },
          "blockage_detail" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ],
            "title" : "Blockage Detail"
          },
          "created_at" : {
            "type" : "string",
            "format" : "date-time",
            "title" : "Created At"
          },
          "updated_at" : {
            "type" : "string",
            "format" : "date-time",
            "title" : "Updated At"
          },
          "playback_url" : {
            "type" : "string",
            "title" : "Playback Url"
          },
          "summary" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/ObservationPackResponse"
            }, {
              "$ref" : "#/components/schemas/VideoRecordSummaryResponse"
            }, {
              "type" : "null"
            } ]
          },
          "error_code" : {
            "anyOf" : [ {
              "type" : "string",
              "enum" : [ "gemini_timeout", "gemini_parse_error", "unsupported_media", "max_attempts_exceeded" ]
            }, {
              "type" : "null"
            } ],
            "title" : "Error Code"
          }
        },
        "required" : [ "blockage_kind", "character_context", "created_at", "goal", "playback_url", "session_id", "situation", "status", "sub_branch", "updated_at" ],
        "title" : "PracticeSessionDetail"
      },
      "PracticeSessionStatusResponse" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "status" : {
            "type" : "string",
            "enum" : [ "analyzing", "analyzed", "failed" ],
            "title" : "Status"
          },
          "error_code" : {
            "anyOf" : [ {
              "type" : "string",
              "enum" : [ "gemini_timeout", "gemini_parse_error", "unsupported_media", "max_attempts_exceeded" ]
            }, {
              "type" : "null"
            } ],
            "title" : "Error Code"
          }
        },
        "required" : [ "status" ],
        "title" : "PracticeSessionStatusResponse"
      },
...
      "ObservationItem" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "start_ms" : {
            "type" : "integer",
            "title" : "Start Ms"
          },
          "end_ms" : {
            "type" : "integer",
            "title" : "End Ms"
          },
          "label" : {
            "type" : "string",
            "title" : "Label"
          },
          "confidence" : {
            "type" : "number",
            "title" : "Confidence"
          }
        },
        "required" : [ "confidence", "end_ms", "label", "start_ms" ],
        "title" : "ObservationItem"
      },
      "ObservationPackResponse" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "summary_id" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Summary Id"
          },
          "observations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ObservationItem"
            },
            "title" : "Observations"
          },
          "uncertainties" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "title" : "Uncertainties"
          }
        },
        "required" : [ "observations", "summary_id", "uncertainties" ],
        "title" : "ObservationPackResponse"
      },
...
    },
    "securitySchemes" : {
      "HTTPBearer" : {
        "type" : "http",
        "scheme" : "bearer"
      }
    }
  }
}